Lots of changes to get things to compile... Will definitely need to actually iterate on things some more, but I needed to do lots of refactoring / fixing tons of bugs in the code :/ I really wanted to avoid having everything in the header file just as that complicates things but it turns out that once again that was the easiet thing to get things working again... Obviously this also required some additional changes elsewhere so that system_driver.hpp wasn't dragging everything from the lightup header into its header and then polluting other pieces of code. I will also probably need to convert more things from std::array and double[] to mfem::Vectors just to be able to move things over to the gpu.
